Add support for setValueImmediate
This new mobile: command restores the ability to manually set the value
of all objects, including UITextField using the direct setValue() JS in
UIAutomation.

By default, POST /value will generate keypress events in a UITextView,
as is in line with the webdriver specification.  With this additon, you
can now use the 'old' appium method of setting directly with setValue
by using:
   mobile:setvalue  [{ element: elementID, value: value }]