Fix fast reset race condition
adb shell am commands return immediately, not when they're finished. am
instrument will close the app once it completes. It's possible for instrument to
be running, the am start command is issued, then instrument finishes and closes
the app. Appium is now waiting for an activity that will never appear and the
tests fail.

The fix is to only start the app once reset has closed the app.

Fixes #662