Added a settings file to instruments that it can read in for "sudo environment variables".
Structured as key=value pairs, it will ignore lines that don't have a '=' in them, and throw
and error if the split on '=' is more then two.

The only thing using it in this pull request is the location of node.

This pull request started as https://github.com/appium/appium/pull/757, but I added the settings file.

Changed sysExec to look at .bash_profile first, and failing that, look at .bashrc

Now only calls /bin/test to check if files exist.

Added more logging to make it clear when some options where read in from the settings file.