Remove setting sessionOverride to false
Remove setting sessionOverride to false when cleaning up the session.
This parameter should left as it is.
Users may specify this value to true in the args when starting Appium server. However, once a session is cleared, this parameter would be as it was, instead of setting it to false.