Fix scroll into view
If the desired object already exists on screen, then avoid scrolling to it.
Without this check, the server will error because there may not be a scrollable.

Scrollable existance depends on screen size. For testing, it's better to have
code that works regardless if scrolling is required or not. This matches the
complex_find behavior.