Re-install Selendroid server only on file changes.
Currently, we look at the timestamp (mtime) of the Selendroid server
file to determine if we need to re-install it. Since we are creating it
every time, this means that we re-install it every single time!
This fix checks the MD5 hash of the server file, so that we re-install
only when the binary is different.