Handle `node-idevice` callback parameter types for `isInstalled`.
This handles #4915.

Node package `node-idevice` has a method on the `IDevice` class
that checks to see if a given app is installed: `isInstalled`.
This method passes a boolean value as the second argument in the
callback function, but `isAppInstalled` expects an array.

The array support appears to be necessary for some other opaque
types so it cannot be removed.  Instead, we check the type of the
second parameter.  If it's `boolean` and `true` then we execute
the positive portion of the if block.