Fix #5408: Prep device and set actual caps during Android noLaunchSetup.
When setting the 'autoLaunch' capability to false, the device was not
prepared and actual capabilities was not set.  Therefore, the driver was
not able to retrieve the device ID or install/remove apps prior to
invoking launchApp.

Add prepareDevice and setActualCapabilities to noLaunchSetup.  Add check
to prepareDevice for initialized ADB Logcat to prevent start (launchApp)
from attempting to prepare the device a second time when using
'autoLaunch=false'.