Allow Java version parsing to match OpenJDK
When an Android device is connected to Appium, the `android-common`
class attempts to get the version of Java installed on the system by
calling `java -version`. If a user has an alternative version of the JDK
installed on their machine (e.g. OpenJDK), the Java version parsing in
android-common will fail due to the RegExp expecting 'java version' as
opposed to what will be output ('openjdk version').

This pull request solves the issue by matching on java and openjdk in
the RegExp matcher inside `androidCommon.parseJavaVersion`.