Specify minimum and maximum Node versions
The "engines" variable in package.json should specify the minimum and the maximum NodeJS version that Appium supports. This will allow cloud providers to check what the maximum supported version is and bundle that as the binary when doing Appium releases.

Note that the engines variable doesn't actually stop users from using Appium for out-of-range Node versions.