chore: fix lerna publish
hopefully.

here's what I did:

1. ran `lerna link convert` which converts all packages which reference _other_ packages to `file:..` urls and adds the same kind of urls to these packages as "dependencies" in the root monorepo.  the upshot here is that we no longer need `lerna bootstrap` because `npm install` from the root will get everything.
2. update packages whose `bin` or `main` fields reference a build artifact; these now use a file in the root which just re-exports whatever the built file was.  otherwise, npm would refuse to install these packages w/o having built the projects first... which would put a monkeywrench in 1.
3. remove `lerna bootstrap` from `postinstall`

so, now, _NO_ `package-lock.json` files should ever be generated in `packages/*`; only the root one.

and hopefully, `lerna publish` will figure out the names of the dependencies (converting `file:` urls to the actual package name) when it's time to do that.  that'd be pretty broken if it didn't!