chore(eslint-config-appium): upgrade to eslint v8 [WIP]
`npm` may or may not successfully install!

- Upgraded to `eslint@8`
- `eslint-plugin-promise` still does not have `eslint@8` support, so this should not land until it does.
- remove unused scripts
- modify peer dependencies to accept a range of versions, which _should_ mean that we only need to update `dependencies` in root `package.json` when upgrading eslint-adjacent stuff in the future _instead_ of in both places--or at least until another major.  I think this is OK, as pinning the peer deps may be quite annoying next round of upgrades
- remove unused `@babel/core`, `eslint-find-rules`

FYI, `@babel/eslint-parser` is still required given we seem to be doing some things that only Babel's parser understands.  I am not sure if this is "bad" or "tolerable", but it may become an issue if we ever choose to abandon babel in the future.