fix(base-driver): supportedLogTypes does not get overwritten
Closes #16738

There's a difference between assigning things to a prototype and public class fields, evidently.

To test this would be best done via interaction between drivers and `base-driver`, but affected drivers do not necessarily run the reusable test suite.  Fundamentally, a unit test for this would be testing Babel.  Unsure of best way to add an E2E test

- Also fiddled with some types, which still aren't quite right, but better.
- Simplified a `_.reduce()` call