fix(appium): remove references to js in typescript code
I think, perhaps, having `.ts` files hanging around so that we can make declarations out of them is bad news.  I got this idea from wdio.  Anyway, the TS alias `appium` (often used as `appium/types`) points to the `types` dir -- not the `build/types` dir. It can't point to that, because recursion. That means an import from `../lib/constants` imports from the _ESM_ `../lib/constants`, which apparently is not understandable by the compiler since there is no `.d.ts` flie living there (even if `constants.js` is typed via docstrings, which it is).  I don't know.  It was easier just to do this for now.  At some point later I'll look at using plain `.d.ts` files for this sort of thing, since it seems like it should work.