fix(support): use legacy peer deps for npm v7+
This partially addresses issue #17287

npm v6 still seems confused, but I think I just need to do further testing on it.  we don't support npm v6 for a dev env so I'm not sure if I should worry about it, but need to determine if the weirdness is more widespread.