feat(typedoc-plugin-appium): typedoc plugin larvae
The first bit of a typedoc plugin for Appium to support `@appiumCommand`, etc.

There was some sort of misunderstanding between Babel and TypeScript which resulted in parts of TypeDoc's API being unavailable.  I did not want to spend a lot of time trying to figure it out, so I just wrote the plugin in TypeScript, since that seems to work.

Right now it's able to grok `executeMethodMap`, `newMethodMap`, and the bigass `METHOD_MAP` in `@appium/base-driver`.  It does not yet deal with tags nor does it generate any documentation, and in fact the current main feature is that it causes `typedoc` to exit before generating any documentation.