chore(appium): add types for cli subcommands
This PR modifies the `Args` type to understand extension subcommands (e.g., `driver install`).

Now the `Args` object that gets passed around will know if the args are for running the server or an extension command, and if the latter, which subcommand and what extra options that subcommand supports (e.g., `driver list --show-installed`)

Further, the `main()` export from `main.js` now uses a conditional type based on the command to know if it will resolve with an `AppiumServer` or `void`.  Previously, this was just a union type; now it is smarter and easier to use.