chore(renovate): update semantic commit behavior
The current intended behavior (to use the "parent dir") as the scope is often or always broken.  An empty scope within parens is actually an invalid conventional commit message.  So instead of `chore():` it will be `chore:`.

We could instead pick a scope (like `deps` or something) but figured it was not important enough to do.