fix(appium): ensure migration always gets applied
The manifest file was not getting its migrations applied _to disk_ in all situations.  Essentially: the migration could repeat over and over, depending on the environment--the changes would only exist in memory.

Added a test, as well.  I didn't originally add a test because the whole migration thing came out of a rabbit hole intended to fix a separate issue.  My bad!