feat(typedoc-plugin-appium): add plugin support
Most of this "just works", but the additional parameters to a plugin command need special handling.  To do that, we have to understand what is a plugin and what is not.

Note that this is incompatible with the "mixin" pattern as found in `BaseDriver` and `FakeDriver`. In other words, a _Plugin_ should not use mixin classes.  I would not recommend mixin classes going forward due to its complexity and opacity.