chore(appium): do not attempt automatic install of extensions in dev env
It sucks to run the "autoinstall extensions" script in CI or the monorepo due to the order in which lifecycle scripts are executed: the `postinstall` script of `appium` will always run _before_ the `prepare` script in the monorepo root.  This completely disables the behavior if Appium has not yet been built.