docs(appium): build-docs now uses `appium-docs`' API
While we _could_ update `appium-docs` to support multiple languages, it's easier to just reuse what we have.

This changes the `build-docs` script in `appium/docs/scripts` to use new APIs `deploy()`, `buildReferenceDocs()` and `deploy()` from `@appium/docutils`.

The "default" language is `en`, and `packages/appium/docs/reference/en` is where TypeDoc outputs its markdown files. This is configured in the `typedoc.json`.  Once TypeDoc has run, each language's `mkdocs.yml` is updated with the new `nav` as computed from the output of TypeDocs.  Importantly--and I don't know if this is right--those TypeDoc-docs are _not_ copied into the specific language dir (if not `en`) and instead are referenced within `en`.  This _may_ pose a navigation problem as it's _might_ be possible for a browsing user to accidentally switch away from the target language.

The script `copy-assets` remains the same and script `build-reference` was removed due to duplication.