feat(typedoc-appium-converter): add packageTitles option and populate extension titles
This changes how packages get their titles (the h1 in the commands page, which is used by mkdocs in the nav bar).

A new option, `packageTitles` is available in `typedoc.json`.  This is an array of objects like so:

```json
{
  "packageTitles": [
    {"name": "@appium/fake-driver": "title": "Fake Driver"},
    {"name": "@appium/fake-plugin": "title": "Fake Plugin"}
  ]
}
```

This allows granular control over how each package is titled.  However, if this option is not provided, the plugin will "extract" a title from the `package.json` of each package, assuming that package is an Appium extension.  By default, it will name drivers `Driver: ${appium.driverName}` and plugins `Plugin: ${appium.pluginName}` where `appium` is a prop in `package.json`.

If the package is _not_ an extension but _does_ have commands -- and currently the only package that fulfills this criteria is `@appium/base-driver` -- _and_ is in a multi-package project, the title will be `${name} Commands` where `name` is from `package.json`.  Otherwise, if we are documenting a project with a single package (e.g., a 3p extension) _and_ we cannot otherwise derive a title, we will just use `Commands` for the title.